fix jl_fs_write type sig
uv_fs_write:offset is int64_t, not size_t.

Closes nolta/Winston.jl#182.

backport of c6bdbfebd5eba9889b9138a7294319da0ba2c526
Conflicts:
	base/fs.jl
	src/jl_uv.c

+++ src/jl_uv.c
@@ -442,7 +442,7 @@ DLLEXPORT int jl_fs_symlink(char *path, char *new_path, int flags)
     return ret;
 }
 
-DLLEXPORT int jl_fs_write(int handle, char *data, size_t len, size_t offset)
+DLLEXPORT int jl_fs_write(int handle, char *data, size_t len, int64_t offset)
 {
     uv_fs_t req;
     uv_buf_t buf[1];
